requirements.txt
#django
Django==4.1.1
djangorestframework==3.13.1


